refactor(api): cleaned up push and pull commands. renamed auth.js to api.js. removed unused prompt file.
