+++ src/commands/push.js
@@ -1,16 +1,16 @@
 const { Command } = require('@oclif/command')
+const { cli } = require('cli-ux')
 const fs = require('fs.promised')
-const fetch = require('node-fetch')
 
-const { DEV_API_ENDPOINT } = require('../constants')
+const { push } = require('../auth')
 
 class PushCommand extends Command {
   async run() {
     const { args } = this.parse(PushCommand)
     const schemaFilePath = args.filePath
     const projectId = args.projectId
-
     let schemaFile
+
     try {
       schemaFile = fs.readFileSync(schemaFilePath).toString()
     } catch (err) {
@@ -20,26 +20,10 @@ class PushCommand extends Command {
       this.log(err)
       return
     }
-    fetch(`${DEV_API_ENDPOINT}/schema/${projectId}`, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        schema: schemaFile
-      })
-    })
-      .then(async res => {
-        if (res.status === 201) {
-          console.log('Success!')
-          return
-        }
-        console.log('Error: Schema not found based on project ID')
-        console.log('Error: Unable to push your local schema. Try again.')
-      })
-      .catch(() => {
-        console.log(
-          'ERROR: There was a server error, please try again or contact us!'
-        )
-      })
+
+    cli.action.start('Saving project shcema...')
+    await push(projectId, schemaFile)
+    cli.action.stop()
   }
 }
